Report header added; Moved eligibility-determining code
